+++ packages/ibm-products-web-components/vite-env.d.ts
@@ -1,8 +1,7 @@
+/// <reference types="vite/client" />
 /**
- * Copyright IBM Corp. 2024, 2024
+ * Copyright IBM Corp. 2026
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
-module.exports = require('babel-preset-ibm-cloud-cognitive')();